chores(server): nitpicks and dedicated abstracted files
