+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-action-v2.ts
@@ -1,11 +1,12 @@
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import { FromTo } from 'src/engine/workspace-manager/workspace-migration-v2/types/from-to.type';
-import { ObjectMetadataUniqueIdentifier } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-action-common-v2';
+import { WorkspaceMigrationObjectWithoutFields } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-input';
 
-type ObjectActionCommon = ObjectMetadataUniqueIdentifier;
+type ObjectActionCommon = {
+  objectMetadataInput: WorkspaceMigrationObjectWithoutFields;
+};
 export type CreateObjectAction = {
   type: 'create_object';
-  object: ObjectMetadataEntity;
 } & ObjectActionCommon;
 
 export type UpdateObjectAction = {
