+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-object-actions.ts
@@ -1,4 +1,3 @@
-import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
 import {
   CreateObjectAction,
   DeleteObjectAction,
@@ -6,16 +5,15 @@ import {
 import { WorkspaceMigrationObjectInput } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-input';
 
 export const getWorkspaceMigrationV2ObjectCreateAction = (
-  input: WorkspaceMigrationObjectInput,
+  objectMetadataInput: WorkspaceMigrationObjectInput,
 ): CreateObjectAction => ({
   type: 'create_object',
-  objectMetadataUniqueIdentifier: input.uniqueIdentifier,
-  object: input as unknown as ObjectMetadataEntity, // TODO prastoin
+  objectMetadataInput,
 });
 
 export const getWorkspaceMigrationV2ObjectDeleteAction = (
-  input: WorkspaceMigrationObjectInput,
+  objectMetadataInput: WorkspaceMigrationObjectInput,
 ): DeleteObjectAction => ({
   type: 'delete_object',
-  objectMetadataUniqueIdentifier: input.uniqueIdentifier,
+  objectMetadataInput,
 });
